Refactor status creation in seed.ts
- Refactored the code to use the upsert method instead of createMany for creating statuses.
- Updated the data structure for creating a new status with associated todos.
- Added new todos for the "Ready" status.
- Created "In Progress" and "Done" statuses if they don't exist.

This commit refactors the code responsible for creating statuses and associated todos in the seed.ts file. The previous implementation used the createMany method, but it has been replaced with upsert to handle both creation and updating of existing records. The data structure for creating a new status has also been updated to include an array of todos. Additionally, this commit adds a new todo under the "Ready" status. If the "Ready" status already exists, it will be updated accordingly. Finally, if either "In Progress" or "Done" statuses do not exist, they will be created.